todo_app/requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
